requirements.txt
numpy~=1.18.5
torchvision~=0.10.0+cu102
pillow~=7.2.0
torch~=1.9.0+cu102
matplotlib~=3.2.2
tqdm~=4.47.0
scikit-image>=0.16.2
opencv-python>=4.5.2.54